+++ src/types/domain.d.ts
@@ -1,0 +1,5 @@
+import { Category } from '.';
+
+type AllCategory = '전체';
+
+export type CategoryOptions = AllCategory | Category;
