+++ src/types/modal.d.ts
@@ -1,0 +1,4 @@
+import RestaurantAddForm from '../components/RestaurantAddForm';
+import RestaurantInformation from '../components/RestaurantInformation';
+
+export type ModalContent = typeof RestaurantAddForm | typeof RestaurantInformation;
